In-Game Debug Console Added, Networking + input bugs fixed
Press space to activate in game debugging console. Fixed bug making
units not deselect when clicking on map.

Found out scripts are completely local for each instance so unit.cs is
invisible to all but owner. This really . . really . . sucks. There is
no easy way to make it considerred a reference to owner's script.